Move some non-sprockets code to tilt/opal, add specs
