+++ lib/opal/sprockets/processor.rb
@@ -8,38 +8,12 @@ require 'opal/sprockets/source_map_server'
 $OPAL_SOURCE_MAPS = {}
 
 module Opal
-  # The Processor class is used to make ruby files (with rb or opal extensions)
-  # available to any sprockets based server. Processor will then get passed any
-  # ruby source file to build. There are some options you can override globally
-  # which effect how certain ruby features are handled:
-  #
-  #   * method_missing_enabled      [true by default]
-  #   * arity_check_enabled         [false by default]
-  #   * const_missing_enabled       [true by default]
-  #   * dynamic_require_severity    [:error by default]
-  #   * source_map_enabled          [true by default]
-  #   * irb_enabled                 [false by default]
-  #   * inline_operators_enabled    [false by default]
-  #
   class Processor
     class << self
-      attr_accessor :method_missing_enabled
-      attr_accessor :arity_check_enabled
-      attr_accessor :const_missing_enabled
-      attr_accessor :dynamic_require_severity
       attr_accessor :source_map_enabled
-      attr_accessor :irb_enabled
-      attr_accessor :inline_operators_enabled
     end
 
-    self.method_missing_enabled      = true
-    self.arity_check_enabled         = false
-    self.const_missing_enabled       = true
-    self.dynamic_require_severity    = :error # :error, :warning or :ignore
     self.source_map_enabled          = true
-    self.irb_enabled                 = false
-    self.inline_operators_enabled    = true
-
 
     def evaluate(context, locals, &block)
       return super unless context.is_a? ::Sprockets::Context
@@ -82,12 +56,6 @@ module Opal
 
     def sprockets_extnames_regexp
       @sprockets_extnames_regexp ||= self.class.sprockets_extnames_regexp(@sprockets)
-    end
-
-    def compiler_options
-      # Not using self.class because otherwise would check subclasses for
-      # attr_accessors they have but are not set.
-      ::Opal::Processor.compiler_options
     end
 
     def process_requires(requires, context)
@@ -200,18 +168,6 @@ module Opal
 
     def stubbed_files
       self.class.stubbed_files
-    end
-
-    def self.compiler_options
-      {
-        :method_missing           => method_missing_enabled,
-        :arity_check              => arity_check_enabled,
-        :const_missing            => const_missing_enabled,
-        :dynamic_require_severity => dynamic_require_severity,
-        :irb                      => irb_enabled,
-        :inline_operators         => inline_operators_enabled,
-        :requirable               => true,
-      }
     end
   end
 end
